requirements.txt
python-telegram-bot==20.4
python-dotenv==1.0.0
flask==2.3.2
pymongo==4.6.1